update control-plane automation

+++ assets/loadtest/control-plane/policies/delete-policies.sh
@@ -10,8 +10,10 @@ s3_policy_arn="arn:aws:iam::${ACCOUNT_ID}:policy/${CLUSTER_NAME}-s3"
 
 route53_policy_arn="arn:aws:iam::${ACCOUNT_ID}:policy/${CLUSTER_NAME}-route53"
 
-aws iam delete-policy \
-    --policy-arn "$dynamo_policy_arn"
+if [[ "$TELEPORT_BACKEND" == "dynamo" ]]; then
+    aws iam delete-policy \
+        --policy-arn "$dynamo_policy_arn"
+fi
 
 aws iam delete-policy \
     --policy-arn "$s3_policy_arn"
